refactor(pty): drop redundant HWND cast in host_has_real_console
windows-sys 0.52 already defines HWND as isize, so the `as isize` cast was
redundant (clippy::unnecessary_cast). Compare directly; parenthesize the
unsafe block so it parses as an expression rather than a statement.

+++ src/NovaTerminal.App/native/src/lib.rs
@@ -28,7 +28,7 @@ mod win32 {
     /// a console. Used to avoid the PSEUDOCONSOLE_PASSTHROUGH path, which drops child
     /// stdout when no real console is present.
     pub fn host_has_real_console() -> bool {
-        (unsafe { GetConsoleWindow() } as isize) != 0
+        (unsafe { GetConsoleWindow() }) != 0
     }
 
     pub fn spawn_with_passthrough(
